+++ src/main/java/cn/foodtower/util/render/Stencil.java
@@ -14,23 +14,23 @@ public class Stencil {
 
     public static void write(boolean renderClipLayer) {
         Stencil.checkSetupFBO();
-        GL11.glClearStencil((int)0);
+        GL11.glClearStencil(0);
         GL11.glClear(GL11.GL_STENCIL_BUFFER_BIT);
         GL11.glEnable(GL11.GL_STENCIL_TEST);
-        GL11.glStencilFunc(GL11.GL_ALWAYS, (int)1, (int)65535);
+        GL11.glStencilFunc(GL11.GL_ALWAYS, 1, 65535);
         GL11.glStencilOp(GL11.GL_KEEP, GL11.GL_KEEP, GL11.GL_REPLACE);
-        if(!renderClipLayer){
+        if (!renderClipLayer) {
             GlStateManager.colorMask(false, false, false, false);
         }
     }
 
     public static void erase(boolean invert) {
-        GL11.glStencilFunc(invert ? GL11.GL_EQUAL : GL11.GL_NOTEQUAL, (int)1, (int)65535);
+        GL11.glStencilFunc(invert ? GL11.GL_EQUAL : GL11.GL_NOTEQUAL, 1, 65535);
         GL11.glStencilOp(GL11.GL_KEEP, GL11.GL_KEEP, GL11.GL_REPLACE);
         GlStateManager.colorMask(true, true, true, true);
         GlStateManager.enableAlpha();
         GlStateManager.enableBlend();
-        GL11.glAlphaFunc( GL11.GL_GREATER, 0.0f );
+        GL11.glAlphaFunc(GL11.GL_GREATER, 0.0f);
     }
 
 
@@ -50,12 +50,12 @@ public class Stencil {
 
 
     public static void setupFBO(Framebuffer fbo) {
-        EXTFramebufferObject.glDeleteRenderbuffersEXT((int)fbo.depthBuffer);
+        EXTFramebufferObject.glDeleteRenderbuffersEXT(fbo.depthBuffer);
         int stencil_depth_buffer_ID = EXTFramebufferObject.glGenRenderbuffersEXT();
-        EXTFramebufferObject.glBindRenderbufferEXT((int)36161, (int)stencil_depth_buffer_ID);
-        EXTFramebufferObject.glRenderbufferStorageEXT((int)36161, (int)34041, (int)Minecraft.getMinecraft().displayWidth, (int)Minecraft.getMinecraft().displayHeight);
-        EXTFramebufferObject.glFramebufferRenderbufferEXT((int)36160, (int)36128, (int)36161, (int)stencil_depth_buffer_ID);
-        EXTFramebufferObject.glFramebufferRenderbufferEXT((int)36160, (int)36096, (int)36161, (int)stencil_depth_buffer_ID);
+        EXTFramebufferObject.glBindRenderbufferEXT(36161, stencil_depth_buffer_ID);
+        EXTFramebufferObject.glRenderbufferStorageEXT(36161, 34041, Minecraft.getMinecraft().displayWidth, Minecraft.getMinecraft().displayHeight);
+        EXTFramebufferObject.glFramebufferRenderbufferEXT(36160, 36128, 36161, stencil_depth_buffer_ID);
+        EXTFramebufferObject.glFramebufferRenderbufferEXT(36160, 36096, 36161, stencil_depth_buffer_ID);
     }
 }
